requirements.txt
Flask==1.0.2
uWSGI==2.0.17.1
PyYAML==5.1.2
requests